Added InterdimensionalPresence
The interdimensional presence will follow the player's exact location in the overworld when the player is in the upside down. This is to load the chunks in the overworld for the CopyChunkMixin to later copy. This will also be used to flicker light on and off later on.

One issue persists. This presence entity seems to get unloaded even though force loaders have been set on the chunk it is in.