Worldgen with ticket delays
Ticket delays doesnt seem to work. Even though tickets created in the upside down are have a two second delay it cant seem to find any chunks in the overworld.

Either find a different approach for loading the chunks in overworld or create a whole new chunk generator for the upside down.